Rearrange file structures as discussed (moved files to folders)
