+++ server/db/assessments.js
@@ -1,5 +1,5 @@
 const environment = process.env.NODE_ENV || 'development'
-const config = require('../knexfile')[environment]
+const config = require('../../knexfile')[environment]
 const connection = require('knex')(config)
 
 function getAssessments(id, db = connection) {}
